skl-lgbm/inference/requirements.txt
numpy
scikit-learn
google-cloud-storage
pandas
lightgbm
fastapi
starlette
